+++ src/main/java/com/livinglogic/ul4/UL4GetAttr.java
@@ -79,4 +79,49 @@ public interface UL4GetAttr
 		}
 		return CallAST.call(context, attribute, args, kwargs);
 	}
+
+	/**
+	<p>Render the attribute (i.e. method) named {@code key} of this object with
+	the positional arguments {@code args} and the keyword arguments
+	{@code kwargs}.</p>
+
+	<p>This default implementation simply defers to
+	{@link #getAttrUL4(EvaluationContext, String)} to get the object (which will
+	probable return a {@link BoundMethod} or {@link GenericBoundMethod} object)
+	and then render this bound method object.</p>
+
+	<p>However this default implementation can be overwritten to implement the
+	functionality of the method directly which skips creating a bound method
+	object.</p>
+
+	<p>However note that {@link #getAttrUL4(EvaluationContext, String)} must
+	still return a bound method object for the method name for the case where
+	the bound method isn't rendered directly.</p>
+
+	@param context The evaluation context.
+	@param key The name of the method to render.
+	@param args Position arguments for the method call.
+	@param kwargs Keyword arguments for the method call.
+	**/
+	default void renderAttrUL4(EvaluationContext context, String key, List<Object> args, Map<String, Object> kwargs)
+	{
+		Object attribute;
+
+		try
+		{
+			attribute = getAttrUL4(context, key);
+		}
+		catch (AttributeException exc)
+		{
+			if (exc.getObject() == this)
+				attribute = new UndefinedAttribute(this, key);
+			else
+				// The {@code AttributeException} originated from another object
+				throw exc;
+		}
+		if (attribute instanceof UL4Render)
+			((UL4Render)attribute).renderUL4(context, args, kwargs);
+		else
+			throw new NotRenderableException(attribute);
+	}
 }